Added filters, enrichments and paginations
Conflicts:

	library/Zend/Service/AgileZen/AgileZen.php
	library/Zend/Service/AgileZen/Resources/Invite.php
	library/Zend/Service/AgileZen/Resources/Story.php

+++ tests/Zend/Service/AgileZen/InviteTest.php
@@ -31,7 +31,7 @@ class InviteTest extends \PHPUnit_Framework_TestCase
         }
         $data = array (
             'email' => constant('TESTS_ZEND_SERVICE_AGILEZEN_ONLINE_INVITE_EMAIL'),
-            'role'  => constant('TESTS_ZEND_SERVICE_AGILEZEN_ONLINE_INVITE_ROLE_ID')
+            'role'  => (integer) constant('TESTS_ZEND_SERVICE_AGILEZEN_ONLINE_INVITE_ROLE_ID')
         );
         $invite = $this->agileZen->addInvite(
                 constant('TESTS_ZEND_SERVICE_AGILEZEN_ONLINE_PROJECT_ID'),
